Initial commit
In this initial commit we have some of the initial jank. A very based C
implementation of communication with Postgres using libpq. This C
version exists because I was having issues with the Odin version and
wanted to make see if it was an Odin issue or a me issue. As I might
expext it was a me issue.

There is also an Odin implementation that queries a local postgres
database with a people table and stuffs it into a dynamic array of
structs.

This is all very manual at the moment, but I believe it might be
possible to use some of Odin's reflection/introspection features to
make a function with a signature similar to the below

pg_query :: proc(query: string, $T: typeid) -> []$T

This function would inspect the result and try to match up the columns
to entries in the struct

I have not played around with yet Odin's reflection/introspection stuff
though so I do not really know how possible any of this is. I'm quite
confident it would be in Jai or Zig, but not really sure for Odin.